propagate change to other projects
