+++ packages/core-modules/test/contracts/modules/CoreUpgradeModule.test.js
@@ -1,12 +1,12 @@
 const hre = require('hardhat');
 const assert = require('assert');
 const assertRevert = require('@synthetixio/core-js/utils/assert-revert');
-const bootstrap = require('../../helpers/bootstrap');
+const { bootstrap } = require('@synthetixio/deployer/utils/tests');
 
 const { ethers } = hre;
 
 describe('CoreUpgradeModule', () => {
-  bootstrap();
+  bootstrap(() => {});
 
   let UpgradeModule, Implementation;
   let owner, user;
